moved components and fixed models
