+++ QuizMakerAngular/src/app/app-routing.module.ts
@@ -7,9 +7,9 @@ import { SignUpComponent } from './components/sign-up/sign-up.component';
 import { DashboardComponent } from './components/dashboard/dashboard.component';
 import { ForgotPasswordComponent } from './components/forgot-password/forgot-password.component';
 import { VerifyEmailComponent } from './components/verify-email/verify-email.component';
-import { UserProfileComponent } from './userProfile/userProfile.component';
-import { UsernameComponent } from './username/username.component';
-import { LandingpageComponent } from './landingpage/landingpage.component';
+import { UserProfileComponent } from './components/userProfile/userProfile.component';
+import { UsernameComponent } from './components/username/username.component';
+import { LandingpageComponent } from './components/landingpage/landingpage.component';
 import { QuizViewComponent } from './components/quiz-view/quiz-view.component';
 import { TeamViewComponent } from './components/team-view/team-view.component';
 import { TeamJoinComponent } from './components/team-join/team-join.component';
