+++ QuizMakerAngular/src/app/components/landingpage/landingpage.component.ts
