+++ QuizMakerAngular/src/app/components/main/main.component.css
