+++ QuizMakerAngular/src/app/components/nav/nav.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { AuthService } from "../services/auth.service";
+import { AuthService } from "../../services/auth.service";
+
 @Component({
   selector: 'app-nav',
   templateUrl: './nav.component.html',
