+++ QuizMakerAngular/src/app/components/password/password.component.html
