+++ QuizMakerAngular/src/app/models/team.ts
@@ -3,4 +3,5 @@ export interface Team {
   teamName: string;
   quizId: number;
   teamScore?: number;
+  captainId?: string;
 }
