+++ QuizMakerAngular/src/app/models/teammember.ts
@@ -1,5 +1,5 @@
 export interface TeamMember {
-  teamId: number;
-  userId: string;
-  captain: boolean;
+  teamId?: number;
+  userId?: string;
+  captain?: boolean;
 }
