+++ QuizMakerAngular/src/app/models/user.ts
@@ -1,6 +1,6 @@
 export interface User {
-  userId: string;
-  email: string;
-  name: string;
-  surname: string;
+  userId?: string;
+  email?: string;
+  name?: string;
+  surname?: string;
 }
